Cast shares to float.

+++ app/models/all_holdings.py
@@ -16,7 +16,7 @@ class AllHoldingsRequest(BaseModel):
 class AllHoldingsRecord(BaseModel):
     ticker: str
     active: bool
-    shares: int
+    shares: float
     price: float
     value: float
     total_return: float
